requirements.txt
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
google-cloud-vision==3.4.1
aiohttp==3.8.4
